pyproject.toml
[project]
name = "ai-plays-pokemon"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
pyboy = {extras = ["pyboy-core"], version = "^2.5.2"}
numpy = "^2.2.5"
pillow = "^11.2.1"
pydantic = "^2.11.3"
burr = "^0.39.1"
loguru = "^0.7.3"
google-genai = "^1.13.0"
dotenv = "^0.9.9"
pydantic-settings = "^2.9.1"
aiofiles = "^24.1.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["./"]
line_length = 100

[tool.black]
line-length = 100

[tool.ruff]
exclude = [
    "./.cache",
    "./.git",
    "./.idea",
    "./.mypy_cache",
    "./.pytest_cache",
    "./.venv",
    "./env",
    "./cached_venv",
    "./.vscode",
    "*migrations*",
]
ignore = [
    "ANN401",
    "D100",
    "D104",
    "D105",
    "D106",
    "D107",
    "D203",
    "D205",
    "D212",
    "D400",
    "D401",
    "D417",
]
line-length = 100
select = [
    "ANN",
    "ASYNC",
    "B",
    "BLE",
    "C4",
    "C9",
    "D",
    "E",
    "ERA",
    "F",
    "ISC",
    "N",
    "PTH",
    "Q",
    "RUF",
    "S",
    "T20",
    "UP",
    "W",
]

[tool.ruff.flake8-quotes]
inline-quotes = "double"

[tool.ruff.mccabe]
max-complexity = 20

[tool.ruff.pep8-naming]
classmethod-decorators = ["classmethod", "validator"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"*/tests/*" = ["S101"]  # ignore assert statements in tests
"conftest.py" = ["S101"]  # ignore assert statements in conftest
"*/prompts.py" = ["E501"]  # ignore line length in prompts

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
